refactor(migration-fix-wizard): 데이터 모델 + RollbackSQLGenerator 분리
대형 파일 정리 7번째(Phase G), 첫 core(비-UI) 대상. migration_fix_wizard.py
(2020줄, 클래스 13개)를 결합도 분석한 결과, Collation 수정 서브시스템
(SmartFixGenerator/CollationFKGraphBuilder/FKSafeCharsetChanger/
BatchFixExecutor/CharsetFixPlanBuilder)은 서로 실제 코드로 얽혀있는
응집된 단위라 이번 패스에서 그대로 두고, 완전히 독립적인 두 부분만 분리.

- migration_fix_models.py (신규): Enum/dataclass 6개(FixStrategy,
  FKDefinition, FixOption, FixWizardStep, FixExecutionResult,
  BatchExecutionResult, CharsetTableInfo) + 공유 순수 헬퍼
  _format_default_sql_clause
- migration_rollback_sql_generator.py (신규): RollbackSQLGenerator
  (다른 5개 로직 클래스와 코드상 교차 참조 없음 확인, 공유 헬퍼만 사용)
- migration_fix_wizard.py: 2020 -> 1495줄 (재수입)

버그 발견/수정: AST의 ClassDef.lineno가 데코레이터 줄을 포함하지 않아서,
추출한 6개 dataclass 전부에서 @dataclass 데코레이터가 누락되는 버그가
있었음 ("BatchExecutionResult() takes no arguments" 등으로 63개 테스트
실패) — 전체 pytest로 즉시 발견, 6곳 모두 데코레이터 추가로 수정.

검증: 관련 pytest 93 passed(데코레이터 수정 전 63 failed였음). 전체
pytest 1809 passed / 1 known-flaky 제외. py_compile 통과. 앱 스모크 성공.
